+++ lib/dealog_backoffice_web/live/settings_live.ex
@@ -2,8 +2,12 @@ defmodule DealogBackofficeWeb.SettingsLive do
   use DealogBackofficeWeb, :live_view
 
   @impl true
-  def mount(_params, _session, socket) do
-    socket = assign(socket, page_title: gettext("Settings"), active_page: :settings)
+  def mount(_params, session, socket) do
+    socket =
+      socket
+      |> assign_defaults(session)
+      |> assign(page_title: gettext("Settings"), active_page: :settings)
+
     {:ok, socket}
   end
 end