+++ lib/dealog_backoffice_web/views/user_confirmation_view.ex
@@ -1,0 +1,3 @@
+defmodule DealogBackofficeWeb.UserConfirmationView do
+  use DealogBackofficeWeb, :view
+end