+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+- `ADD` Mail support for registration, password reset, email change is available
+- `CNG` Update dependencies
+- `ADD` Authentication is implemented for all content pages
+- `ADD` Users can change their email address
+- `ADD` Users can change their password
+- `ADD` Users can authenticate
+- `ADD` Users can register
+
 ## 0.4.0
 
 This release covers message publishing for a single organization.
@@ -101,4 +109,3 @@ This is the initial version setting up the base system with the page structure a
 - `CNG` means changed behavior
 - `FIX` means a bugfix or fix of a glitch
 - `REM` means a removed feature
-